+++ contracts/interfaces/IPriceFeed.sol
@@ -4,10 +4,10 @@ pragma solidity 0.8.6;
 
 interface IPriceFeed {
     // --- Events ---
-    event LastGoodPriceUpdated(uint _lastGoodPrice);
+    event LastGoodPriceUpdated(uint256 _lastGoodPrice);
 
     // --- Function ---
-    // function fetchPrice() external returns (uint);
+    // function fetchPrice() external returns (uint256);
 
-    function fetchPrice_v() external view returns (uint);
+    function fetchPrice_v() external view returns (uint256);
 }
